+++ CustomAbilities/FoniasAbilityComponent.cs
@@ -1,5 +1,24 @@
 ﻿namespace VoiceOfTheCommunity.CustomAbilities;
 
-public sealed class FoniasAbilityComponent : AbilityComponentHack<FoniasAbility>
+public sealed class FoniasAbilityComponent : AbilityComponentHack<FoniasAbility>, IStackable
 {
+    public int currentCount { get; set; }
+
+    public float stack
+    {
+        get
+        {
+            return currentCount;
+        }
+        set
+        {
+            currentCount = (int)value;
+        }
+    }
+
+    public override void Initialize()
+    {
+        base.Initialize();
+        baseAbility.component = this;
+    }
 }