+++ Thunderstore-EN/CHANGELOG.md
@@ -2,7 +2,7 @@
 Let's see if we can pass this time without any bugs!
 
 Changes:
-- 'Monk's Tiger Claw Bracers' now properly enabels auto attack when used with Skul Patcher
+- 'Monk's Tiger Claw Bracers' now properly enables auto attack when used with Skul Patcher
 - 'Fonias' now properly stacks each time killing a boss
 - Re-sprited 'Rusty Chalice' and 'Goddess's Chalice'
 - Re-sprited 'Dwarven Legend's Pickaxe'