+++ Thunderstore-EN/CHANGELOG.md
@@ -1,4 +1,13 @@
-1.1.4 with A change.
+1.1.5 with a few changes!
 
 Changes:
-- 'Tattered Cat Plushie' now has proper line breaks in its description
+- 'Soul Flame Scythe' now has a new description and lore
+- 'Soul Flame Scythe' now increases the chance to gain Soul Fire from enemies by 25% while under the Heirloom inscription effect
+- 'Rusty Shovel' now has a lingering damage increase instead of a one-time use
+- 'Fonias' no longer adds crit chance and damage
+- 'Fonias' now counts your boss kills after the item has been picked up
+- 'Craftsman's Chisel' and 'Master Craftsman's Chisel' are no longer activated by 'Unstable Quintessence'
+- 'Monk's Bracers' no longer disables turbo-attack while it is enabled by Skul Patcher
+- 'Secret of the King' now grants 35% crit damage
+- 'Demoman's Bottle' no longer cancels attacks towards enemies that lack status info
+- 'Forgotten Company Helmet' now has new Korean lore